[Decomp][Misc] Cohesion split: extract instance-difficulty handlers to MiscHandlerInstance.cpp
